+++ TestJava2026/src/main/java/com/springboot/test/adapter/in/controller/PricesController.java
@@ -1,5 +1,9 @@
 package com.springboot.test.adapter.in.controller;
 
+import com.springboot.test.adapter.in.dto.PriceResponse;
+import com.springboot.test.adapter.in.mapper.PriceResponseMapper;
+import com.springboot.test.application.port.in.GetBrandProductPriceUseCase;
+import com.springboot.test.domain.model.Prices;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.format.annotation.DateTimeFormat;
@@ -12,11 +16,21 @@ import java.time.LocalDateTime;
 @RestController
 public class PricesController {
     private Logger logger = LoggerFactory.getLogger(PricesController.class);
+    private final GetBrandProductPriceUseCase getBrandProductPriceUseCase;
+    private final PriceResponseMapper priceResponseMapper;
+
+    public PricesController(
+        GetBrandProductPriceUseCase getBrandProductPriceUseCase,
+        PriceResponseMapper priceResponseMapper
+    ) {
+        this.getBrandProductPriceUseCase = getBrandProductPriceUseCase;
+        this.priceResponseMapper = priceResponseMapper;
+    }
 
     @GetMapping("/prices")
-    public ResponseEntity<?> getPriceAtDate( @RequestParam Integer brandId, @RequestParam Integer productId, @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime date ) {
-        // TODO: implement
+    public ResponseEntity<PriceResponse> getPriceAtDate(@RequestParam Integer brandId, @RequestParam Integer productId, @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime date ) {
         this.logger.debug("[PricesController][getPriceAtDate] DEBUG: brandId={}, productId={}, date={}", brandId, productId, date.toString());
-        return ResponseEntity.ok("TODO");
+        Prices currentPrice = this.getBrandProductPriceUseCase.getBrandProductPrice(brandId, productId, date);
+        return ResponseEntity.ok(this.priceResponseMapper.toResponse(currentPrice));
     }
 }