+++ TestJava2026/src/main/java/com/springboot/test/adapter/in/dto/PriceResponse.java
@@ -1,0 +1,17 @@
+package com.springboot.test.adapter.in.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+public class PriceResponse {
+    private Integer brandId;
+    private Integer productId;
+    private LocalDateTime startDate;
+    private LocalDateTime endDate;
+    private Double price;
+    // TODO: Wouldn't currency make sense too?
+}