+++ TestJava2026/src/main/java/com/springboot/test/adapter/out/repository/PricesRepository.java
@@ -1,8 +1,20 @@
 package com.springboot.test.adapter.out.repository;
 
+import com.springboot.test.adapter.out.entity.PricesEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+import java.time.LocalDateTime;
+import java.util.List;
 
 @Repository
-public class PricesRepository {
-    // TODO: Implement
+public interface PricesRepository extends JpaRepository<PricesEntity, Integer> {
+    @Query(
+        "SELECT p " +
+        "FROM PricesEntity p " +
+        "WHERE " +
+            ":brandId = p.brandId AND" +
+            ":productId = p.productId AND " +
+            "(:date BETWEEN p.startDate AND p.endDate)")
+    List<PricesEntity> findByBrandIdAndProductIdAndDate(Integer brandId, Integer productId, LocalDateTime date);
 }