+++ TestJava2026/src/main/java/com/springboot/test/application/port/out/WritePricesPort.java
@@ -1,0 +1,5 @@
+package com.springboot.test.application.port.out;
+
+public interface WritePricesPort {
+    // No need to save anything, done for reference
+}